//Transactional(readOnly = true) //заинжетил модификатор доступа BCryptPasswordEncoder

+++ src/main/java/ru/kata/spring/boot_security/demo/dao/RoleDao.java
@@ -16,5 +16,5 @@ public interface RoleDao {
 
     public void edit(Role role);
 
-    public Role getById(int id);
+    public Role getById(Long id);
 }
